% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\docType{data}
\name{default_biomart_attributes}
\alias{default_biomart_attributes}
\title{Default biomart attributes to retrieve}
\format{
An object of class \code{character} of length 4.
}
\usage{
default_biomart_attributes
}
\description{
Default biomart attributes to retrieve
}
\keyword{datasets}
