% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{p_makeOrgPackage}
\alias{p_makeOrgPackage}
\title{Copy from AnnotationForge}
\usage{
p_makeOrgPackage(
  data,
  version,
  maintainer,
  author,
  outputDir = getwd(),
  tax_id,
  genus = NULL,
  species = NULL,
  goTable = NA,
  databaseOnly = FALSE,
  verbose = TRUE,
  gid_type = "eg"
)
}
\description{
Copy from AnnotationForge
}
