% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_annotation}
\alias{pull_annotation}
\title{Pull Annotation from BioMart}
\usage{
pull_annotation(
  genes,
  mart_info,
  filter = "ensembl_gene_id",
  dest_attrs = default_biomart_attributes,
  collapse_sep = "|"
)
}
\arguments{
\item{genes}{Ensembl gene id}

\item{mart_info}{A list contains BioMart information:
\code{host}, \code{mart} and \code{dataset}}

\item{filter}{Filter name used to query BioMart}

\item{dest_attrs}{BioMart attributes to pull down}

\item{collapse_sep}{Separator for collapsed attributes}
}
\value{
a tibble
}
\description{
Pull Annotation from BioMart
}
