% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data_from_biomart}
\alias{pull_data_from_biomart}
\title{Pull data table from biomart}
\usage{
pull_data_from_biomart(host, mart, dataset, dest_attrs, filename = NULL)
}
\arguments{
\item{host}{Host url such as plants.ensembl.org}

\item{mart}{Mart name such as plants_mart}

\item{dataset}{Dataset name such as taestivum_eg_gene}

\item{dest_attrs}{Attributes you want to retrieve.
It can be a named vector or named list.}

\item{filename}{File Name to save table.}
}
\description{
Pull data table from biomart
}
