% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{makeOrgDbFromBiomart}
\alias{makeOrgDbFromBiomart}
\title{Make a OrgDb package from annotations available on a BioMart database}
\usage{
makeOrgDbFromBiomart(
  tax_id,
  maintainer,
  author,
  genus = NULL,
  species = NULL,
  version = "0.0.0.9000",
  gid_type = "eg",
  mart = "ENSEMBL_MART_ENSEMBL",
  dataset = "hsapiens_gene_ensembl",
  host = "www.ensembl.org",
  outputDir = getwd()
)
}
\arguments{
\item{tax_id}{The Taxonomy ID that represents your organism.  (NCBI has
  a nice online browser for finding the one you need)}

\item{maintainer}{Who is the package maintainer? (must include email to be 
    valid)}

\item{author}{Who is the creator of this package?}

\item{genus}{Single string indicating the genus.}

\item{species}{Single string indicating the species.}

\item{version}{What is the version number for this package? format: \\'x.y.z\\'}

\item{gid_type}{the type of central identifier.}

\item{mart}{which BioMart database to use. Get the list of all available
BioMart databases with the \code{\link[biomaRt]{listMarts}} function from
the biomaRt package.}

\item{dataset}{which dataset from BioMart.}

\item{host}{The host URL of the BioMart.}

\item{outputDir}{A path where the package source should be assembled.}
}
\description{
Make a OrgDb package from annotations available on a BioMart database
}
\examples{
makeOrgDbFromBiomart(
  mart = "plants_mart",
  dataset = "taestivum_eg_gene",
  host = "https://plants.ensembl.org",
  version = "0.0.0.9000",
  maintainer = "Altair Wei <altair_wei@outlook.com>",
  author = "Altair Wei <altair_wei@outlook.com>",
  outputDir = ".",
  tax_id = "4565",
  genus = "Triticum",
  species = "aestivum",
  gid_type = "iwgsc"
)
}
